# Improve TSI query planning with cached bitsets